Add tests to cover a bug where logs are forwarded in HTTP response headers
PHP-FPM sends logs in the FastCGI output under the key `PHP message:`. Those lines end up in the HTTP headers. They are parsed by Hoa/FastCGI as response headers.

In the end those logs are considered as part as the HTTP response headers but they are not.

PHP-FPM should not write those in the FastCGI connection.