Revert #202 for reasons explained in the PR (copied below)
I just realized that the code example actually replaces all the existing environment variables.

Additionally I'm not sure it's a good idea at all because it is not an atomic deployment: first cloudformation will deploy the function and then (seconds later) it defines the environment variables. For a short period the variables will not exist.

Until we document a proper solution (see #7) I think it's best to revert this bit because it will create confusion and frustration for users. Sorry for not catching that sooner!