Add a new `bref deployment` command to debug failed deployments
If an error occurs when deploying with SAM we need to look into the *Events* tab of the stack in the CloudFormation dashboard.

This works but isn't friendly and practical. This PR introduces the following Bref command to list the logs from the latest deployments and understand what went wrong:

```bash
vendor/bin/bref deployment <stack-name>
```

Only events from the last 24 hours are displayed. We could make it configurable in the future (PRs welcome).