Define explicitly the public API of the Bref codebase
`@internal` let's us signal which classes are not covered by the BC promise. These classes are internal to Bref and can change at any time.

`final` let's us prevent users from extending a class, giving us more freedom to change them without breaking BC.

We are free to remove some limitations like these in the future of course (these are not final decisions), but it's safer to start with limitations and remove them.