try to make the cdn configuration simpler
The serverless-pseudo-parameters plugin is not really needed
and the trade of of a slightly less readable syntax is definitely worth
the 1 fewer dependency and not needing to having npm.

Then I also removed the slightly inconsistent definition of the assets target.
It talked and configured the source partially for an s3 website source
but not fully. It now is just a plain s3origin with a reference to the manual.