Don't kill self
When a previous request times out and doesn't get a change to clean up,
any subsequent request can many times fail.

This is because before starting a php-fpm process we are trying to kill
any existing ones. When process ids get reused between requests and
the current process id is the same as the previously timed out process, we
would be killing self.

When this is detected my change will just do the socket and pid file
cleanups, but, leave the process alone.